Merged in fix_doc_typos (pull request #27)
Fixed typos in documentation

If used as internal plugin, the documentation can be compiled within
hoomd using doxygen, which fails if there are typos in the documentation.
This is not a problem if the plugins are either used as external plugin
or if the hoomd documentation is not compiled (default hoomd behavior).

Approved-by: Michael Howard <mphoward@princeton.edu>